Consolidate z/OS version into windows and bugfix
There is no reason to have to versions of the same files. 
The changes to get it to run on z are trivial, and should be covered by the wiki.
Fixed a bug when working with mocks.
An index belonging to another array was used when moving to GENERATED-SOURCE-STATEMENT, this causes unintended behavior on the Z, when using statements that span multiple lines.