Fix issues with period in comments
This fixes an issue where you could not have a period in a COBOL comment, without the unit test breaking.
This also fixes an issue where you could not use unit test keywords like "expect" in testcase or testsuite description. 
This also provides a better solution to the issue of a missing period at the end of the unit test.