Simplify mock verification, improved handling of . , count can be "zero"
The verification of mocks were overly complex, this caused problems with locating keywords. The keyword was allready parsed as token, so easy simplification. 
Using . as end-if or in SQL, was not possible. . detection improved to only detect lines where . is the only character. 
Added ZERO as a viable count in a verify.